Red: Added expectation for keypath() macro
One problem with KVO is that keypaths are specified as strings.  If you
refactor properties, the Xcode refactoring process won't make changes
to those keypath strings, meaning your code breaks in subtle ways.  The
keypath() macro evaluates the object/property expression as both a
string and as an expression.  The string allows it to be passed into a
specific object type, and the expression ensures that the refactoring
process can identify it and refactor it along with other expressions.